refactor: move GitHub release methods to UpdateService and simplify test page
- Add getLatestReleaseFromGitHub() and getLatestPrereleaseFromGitHub() to UpdateService
- Add ReleaseInfo model class
- Rename _PrereleaseUpdateCard to _UpdateToLatestCard with isPrerelease-aware UI
- Add l10n keys: startUpdate, startUpdatePreview, updateToStable, updateToPreview, etc.
- Remove _UpdateCard (stable release update) and keep only preview update flow
- Add UpdateInfo ChangeNotifier to manage update state per channel
- Add _DownloadProgressState for download dialog status with onStatus callback
