+++ lib/pages/about_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/providers/app_info_provider.dart';
-import 'package:bugaoshan/serivces/update_service.dart';
+import 'package:bugaoshan/services/update_service.dart';
 import 'package:bugaoshan/utils/open_link.dart';
 import 'package:bugaoshan/pages/test_page.dart';
 import 'package:bugaoshan/widgets/dialog/dialog.dart';
