+++ lib/pages/campus/train_program/train_program_provider.dart
@@ -4,7 +4,7 @@ import 'package:flutter/widgets.dart';
 import 'package:bugaoshan/pages/campus/train_program/models/train_program.dart';
 import 'package:bugaoshan/pages/campus/train_program/models/train_program_model.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
-import 'package:bugaoshan/serivces/scu_auth_service.dart';
+import 'package:bugaoshan/services/scu_auth_service.dart';
 import 'package:bugaoshan/utils/session_expiry_handler.dart';
 
 enum TrainProgramLoadState { idle, loading, loaded, error }
