+++ lib/pages/scu_login_page.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
-import 'package:bugaoshan/serivces/scu_auth_service.dart';
-import 'package:bugaoshan/serivces/ocr_service.dart';
+import 'package:bugaoshan/services/scu_auth_service.dart';
+import 'package:bugaoshan/services/ocr_service.dart';
 
 class ScuLoginPage extends StatefulWidget {
   const ScuLoginPage({super.key});
