+++ lib/services/ccyl_oauth_service.dart
@@ -1,7 +1,7 @@
 import 'package:bugaoshan/injection/injector.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
 import 'package:flutter/widgets.dart';
-import 'package:bugaoshan/serivces/scu_auth_service.dart';
+import 'package:bugaoshan/services/scu_auth_service.dart';
 
 class CcylOAuthService {
   static const _idBase = 'https://id.scu.edu.cn';
