+++ lib/services/ocr_service.dart
