+++ lib/services/scu_auth_service.dart
